requirements.txt
asgiref==3.9.1
blinker==1.9.0
click==8.2.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector==2.2.9
passlib==1.7.4
pbkdf2==1.3
PyMySQL==1.1.2
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
